Select sources if sources are empty
